chore : README 업데이트

+++ src/main/java/com/cheolhyeon/free_commnunity_1/post/controller/response/PostReadResponse.java
@@ -1,6 +1,7 @@
 package com.cheolhyeon.free_commnunity_1.post.controller.response;
 
 
+import com.cheolhyeon.free_commnunity_1.comment.controller.reponse.CommentPageResponse;
 import com.cheolhyeon.free_commnunity_1.comment.controller.reponse.CommentReadResponse;
 import com.cheolhyeon.free_commnunity_1.post.domain.Post;
 import lombok.*;
@@ -20,12 +21,12 @@ public class PostReadResponse {
     private String imageUrl;
     private String categoryName;
     private Long viewCount;
-    private int commentCount;
+    private Long commentCount;
     private Long likeCount;
     private List<CommentReadResponse> comments;
     private LocalDateTime updatedAt;
 
-    public static PostReadResponse from(Post post, Long currentViewCount, String writer, String categoryName, List<CommentReadResponse> comments, Long currentPostLikeCount, int commentCount) {
+    public static PostReadResponse from(Post post, Long currentViewCount, String writer, String categoryName, List<CommentReadResponse> comments, Long currentPostLikeCount, Long commentCount) {
         PostReadResponse response = new PostReadResponse();
         response.title = post.getTitle();
         response.content = post.getContent();
